povsite/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==5.0.3
django-filter==24.2
django-froala-editor==4.2.0
django-modeltranslation==0.18.11
djangorestframework==3.15.1
executing==2.0.1
ipython==8.22.2
jedi==0.19.1
Markdown==3.6
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
style==1.1.0
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
wcwidth==0.2.13
